pyproject.toml
[project]
name = "coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.68.0",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
]
